refactor(root): gate dashboard graph test runtime
Move DashboardGraphTestRuntimeV1 out of the production dashboard
module into a test-transport-only fixture so #1088 test runtimes
do not ship in composition code.
